removing changes made in session 5 see
Removing/deleting the changes I made in Session 5 See guide using GH Desktop so this Feature Branch can be back in sync with Master branch.

Note: Master branch at this moment contains a direct commit I made in the Master branch when James asked Jenn to test out the repo to see if a direct commit in Master would create a pull request. 
The direct commit still went through to make changes in the Master branch, despite the pull request never getting approved for merge in the first place and the spell check action failing to detect the edit I included.